# swagger.yaml
openapi: 3.0.0
info:
  title: Fitness API
  description: |
    API RESTful complète pour l'application mobile Fitness.
    Supporte gestion des utilisateurs, courses, séances, pas, poids, amis, messages et badges.
  version: 1.0.0
  contact:
    name: API Support
    email: support@fitnesssapi.com
  license:
    name: MIT
servers:
  - url: https://fitnesssapi.onrender.com
    description: Production Server
  - url: http://localhost:3000
    description: Local Development Server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentification par Bearer Token (JWT)
  schemas:
...
    Workout:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Upper Body
        workoutType:
          type: string
          example: strength
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        durationMinutes:
          type: integer
        caloriesBurned:
          type: number
        notes:
          type: string
          nullable: true
        exerciseCount:
          type: integer
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        exerciseLogs:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseLog'
...
    WorkoutCreate:
      type: object
      required:
        - name
        - startTime
        - endTime
        - durationMinutes
      properties:
        name:
          type: string
        workoutType:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        durationMinutes:
          type: integer
        caloriesBurned:
          type: number
        notes:
          type: string
        exercises:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseInput'
    WorkoutUpdate:
      type: object
      properties:
        name:
          type: string
        workoutType:
          type: string
        durationMinutes:
          type: integer
        caloriesBurned:
          type: number
        notes:
          type: string
    ExerciseLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        exerciseId:
          type: string
        exerciseName:
          type: string
        exerciseType:
          type: string
        notes:
          type: string
          nullable: true
        setLogs:
          type: array
          items:
            $ref: '#/components/schemas/SetLog'
        cardioLogs:
          type: array
          items:
            $ref: '#/components/schemas/CardioLog'
    ExerciseInput:
      type: object
      required:
        - exerciseId
        - exerciseName
      properties:
        exerciseId:
          type: string
        exerciseName:
          type: string
        exerciseType:
          type: string
        notes:
          type: string
        sets:
          type: array
          items:
            $ref: '#/components/schemas/SetInput'
        cardio:
          $ref: '#/components/schemas/CardioInput'
    SetLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        setNumber:
          type: integer
        reps:
          type: integer
        weightKg:
          type: number
        isCompleted:
          type: boolean
        restTimeSec:
          type: integer
    SetInput:
      type: object
      required:
        - setNumber
      properties:
        setNumber:
          type: integer
        reps:
          type: integer
        weightKg:
          type: number
        isCompleted:
          type: boolean
        restTimeSec:
          type: integer
    CardioLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        durationMinutes:
          type: number
        distanceKm:
          type: number
        avgSpeedKmh:
          type: number
        maxSpeedKmh:
          type: number
        caloriesBurned:
          type: number
        avgHeartRate:
          type: integer
        maxHeartRate:
          type: integer
    CardioInput:
      type: object
      properties:
        durationMinutes:
          type: number
        distanceKm:
          type: number
        avgSpeedKmh:
          type: number
        maxSpeedKmh:
          type: number
        caloriesBurned:
          type: number
        avgHeartRate:
          type: integer
        maxHeartRate:
          type: integer
    WorkoutStats:
      type: object
      properties:
        totalWorkouts:
          type: integer
        totalCaloriesBurned:
          type: number
        workoutsThisWeek:
          type: integer

    # ===== Custom Program Schemas =====
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Code d'erreur
          example: VALIDATION_ERROR
        message:
          type: string
          description: Message d'erreur lisible
          example: Email or password incorrect
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ===== Sync Schemas =====
...
paths:
...
  /workouts/stats/summary:
    get:
      tags:
        - Workouts
      summary: Statistiques des séances
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Statistiques
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutStats'
  /workouts:
    get:
      tags:
        - Workouts
      summary: Lister les séances
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: details
          in: query
          schema:
            type: string
            enum: [full]
      responses:
        '200':
          description: Liste des séances
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'

    post:
      tags:
        - Workouts
      summary: Créer une séance
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutCreate'
      responses:
        '201':
          description: Séance créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workout'
        '400':
          description: Erreur de validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /workouts/{id}:
    get:
      tags:
        - Workouts
      summary: Récupérer une séance par ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Séance trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workout'
        '404':
          description: Séance non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - Workouts
      summary: Mettre à jour une séance
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutUpdate'
      responses:
        '200':
          description: Séance mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workout'
        '404':
          description: Séance non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Workouts
      summary: Supprimer une séance
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Séance supprimée
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Séance non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ===== CUSTOM PROGRAMS =====
...